fixed resolve1 #1
