+++ classification/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for assignment project.
+Django settings for classification project.
 
 Generated by 'django-admin startproject' using Django 1.10.
 
@@ -20,7 +20,7 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 # See https://docs.djangoproject.com/en/1.10/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = '6lgu2l9z4)$4h@-p01_o6#&qkadn8bvgnw!^yl5a)#adj5^rx-'
+SECRET_KEY = 'k3qbeeb8cr0#v)s(b5ac3sf@(6s3uk!-p1#djd^4jl_#blb35+'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -31,7 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'trial.apps.TrialConfig',
+    'showcategory',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'assignment.urls'
+ROOT_URLCONF = 'classification.urls'
 
 TEMPLATES = [
     {
@@ -68,7 +68,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'assignment.wsgi.application'
+WSGI_APPLICATION = 'classification.wsgi.application'
 
 
 # Database
@@ -106,7 +106,7 @@ AUTH_PASSWORD_VALIDATORS = [
 
 LANGUAGE_CODE = 'en-us'
 
-TIME_ZONE = 'Asia/Tokyo'
+TIME_ZONE = 'UTC'
 
 USE_I18N = True
