+++ classification/urls.py
@@ -1,4 +1,4 @@
-"""assignment URL Configuration
+"""classification URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.10/topics/http/urls/
@@ -17,6 +17,6 @@ from django.conf.urls import include, url
 from django.contrib import admin
 
 urlpatterns = [
-    url(r'^trial/', include('trial.urls')),
+    url(r'^showcategory/', include('showcategory.urls')),
     url(r'^admin/', admin.site.urls),
 ]
