--- machine_leaning_method_app/.cache/v/cache/lastfailed
@@ -1,3 +1,0 @@
-{
-  "urls.py": true
-}
