--- machine_leaning_method_app/.cache/v/pep8/mtimes
@@ -1,14 +1,0 @@
-{
-  "/Users/Masashi/Desktop/Python/assignment/trial/get_test_data.py": [
-    1473310481.0,
-    []
-  ],
-  "/Users/Masashi/Desktop/Python/assignment/trial/ngwords.py": [
-    1473310771.0,
-    []
-  ],
-  "/Users/Masashi/Desktop/Python/assignment/trial/train_classify.py": [
-    1473310424.0,
-    []
-  ]
-}
