--- machine_leaning_method_app/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TrialConfig(AppConfig):
-    name = 'trial'
