--- machine_leaning_method_app/management/commands/.cache/v/cache/lastfailed
@@ -1,1 +1,0 @@
-{}
