--- machine_leaning_method_app/management/commands/.cache/v/pep8/mtimes
@@ -1,6 +1,0 @@
-{
-  "/Users/Masashi/Desktop/Python/assignment/trial/management/commands/get_train_data_from_sites.py": [
-    1473310900.0,
-    []
-  ]
-}
