+++ showcategory/admin.py
@@ -1,6 +1,3 @@
 from django.contrib import admin
 
 # Register your models here.
-from .models import Category
-
-admin.site.register(Category)
