+++ showcategory/get_train_data.py
@@ -1,11 +1,12 @@
-import trial.testwords
+import showcategory.testwords
 # 以下カテゴリーごとに得られた単語をテキストファイルで読み込む
 categories = ["entertainment", "sports", "fun", "domestic",
               "abroad", "column", "it_science", "gourmet"]
 dict = {}
 vocabulary = []
 for i in range(1, 9):
-    f = open("trial/testwords/testwords_category" + str(i) + ".txt", "r")
+    f = open("showcategory/testwords/testwords_category" +
+             str(i) + ".txt", "r")
     words = f.readlines()
     wordslist = []
     for word in words:
