+++ showcategory/management/commands/get_train_data_from_sites.py
@@ -1,17 +1,15 @@
 # -*- coding: utf-8 -*-
 
 from django.core.management.base import BaseCommand
-from getmostfrequentwords import *
+from showcategory.management.commands.getmostfrequentwords import *
 import urllib.request
 from bs4 import BeautifulSoup
-from urllib.parse import urljoin
-import MeCab
-import collections
+
 
 class Command(BaseCommand):
 
     def handle(self, *args, **options):
-        
+
         # 以下でグノシーから教師データに用いる記事を取得
         url = "https://gunosy.com/"
         response = urllib.request.urlopen(url)
@@ -19,7 +17,8 @@ class Command(BaseCommand):
         soup = BeautifulSoup(html, "html.parser")
         for i in range(1, 9):
             # for i in range(1,2):
-            f = open('trial/testwords/article_category' + str(i) + '.txt', 'w')
+            f = open(
+                'showcategory/testwords/article_category' + str(i) + '.txt', 'w')
             categories_link = soup.find(class_='nav_color_' + str(i)).find('a')
             categories_url = str(categories_link.get('href'))
             print(categories_url)
@@ -30,9 +29,6 @@ class Command(BaseCommand):
                 # for j in range(2,3):
                 article_links = soup.find(
                     class_='main').find_all(class_='list_title')
-                article_list_page = str(
-                    soup.find(class_='pager-link-option')
-                    .find('a').get('href'))
                 for article_link in article_links:
                     article_url = str(article_link.find('a').attrs['href'])
                     response = urllib.request.urlopen(article_url)
@@ -42,8 +38,7 @@ class Command(BaseCommand):
                         class_='article gtm-click').find_all('p')
                     print(article_url)
                     for article in article_contents:
-                        article_content = article.get_text()
-                        f.write(article_content)
+                        f.write(article.get_text())
                     # 記事内容を取得し終わったら次の記事リストページへ
                 response = urllib.request.urlopen(
                     categories_url + '?page=' + str(j))
@@ -54,10 +49,10 @@ class Command(BaseCommand):
         # 以下で教師データを生成
         # for i in range(1,2):
         for i in range(1, 9):
-            f = open("trial/testwords/testwords_category" +
+            f = open("showcategory/testwords/testwords_category" +
                      str(i) + ".txt", "w")
             trainwords = getmostfrequentwords(
-                'trial/testwords/article_category' + str(i) + '.txt')
+                'showcategory/testwords/article_category' + str(i) + '.txt')
             for trainword in trainwords:
                 f.write(str(trainword[0]) + "\n")
             f.close()
